backend/requirements.txt
async-timeout==4.0.3
bcrypt==4.1.3
bidict==0.23.1
blinker==1.8.2
click==8.1.7
Flask==3.0.3
flask-extended==0.2
Flask-JWT-Extended==4.6.0
flask-redis==0.4.0
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
greenlet==3.0.3
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
PyJWT==2.8.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.2
redis==5.0.4
simple-websocket==1.0.0
SQLAlchemy==2.0.30
typing_extensions==4.12.0
uuid==1.30
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.1.2
